Create lib/board.rb Created lib/player.rb
